Refactor messenger to network package, minor cleanup
