+++ bukkit/src/main/java/net/william278/huskhomes/BukkitHuskHomes.java
@@ -20,9 +20,9 @@ import net.william278.huskhomes.event.EventDispatcher;
 import net.william278.huskhomes.hook.*;
 import net.william278.huskhomes.listener.BukkitEventListener;
 import net.william278.huskhomes.listener.EventListener;
-import net.william278.huskhomes.messenger.BukkitPluginMessenger;
-import net.william278.huskhomes.messenger.NetworkMessenger;
-import net.william278.huskhomes.messenger.BukkitRedisMessenger;
+import net.william278.huskhomes.network.BukkitPluginMessenger;
+import net.william278.huskhomes.network.NetworkMessenger;
+import net.william278.huskhomes.network.BukkitRedisMessenger;
 import net.william278.huskhomes.migrator.LegacyMigrator;
 import net.william278.huskhomes.migrator.Migrator;
 import net.william278.huskhomes.player.BukkitPlayer;
