+++ common/src/main/java/net/william278/huskhomes/request/RequestManager.java
@@ -1,8 +1,8 @@
 package net.william278.huskhomes.request;
 
 import net.william278.huskhomes.HuskHomes;
-import net.william278.huskhomes.messenger.Message;
-import net.william278.huskhomes.messenger.MessagePayload;
+import net.william278.huskhomes.network.Message;
+import net.william278.huskhomes.network.MessagePayload;
 import net.william278.huskhomes.player.OnlineUser;
 import net.william278.huskhomes.player.User;
 import net.william278.huskhomes.teleport.Teleport;
